Added two, final corrections and succesfull runs
