+++ ml_project/heart_disease_classification/models/model_fit_predict.py
@@ -6,7 +6,7 @@ import pandas as pd
 from sklearn.compose import ColumnTransformer
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.linear_model import LogisticRegression
-from sklearn.metrics import r2_score, mean_squared_error, mean_absolute_error
+from sklearn.metrics import r2_score, mean_squared_error, mean_absolute_error, accuracy_score
 from sklearn.pipeline import Pipeline
 
 from heart_disease_classification.params.train_params import TrainingParams
@@ -31,23 +31,20 @@ def train_model(
 
 
 def predict_model(
-    model: Pipeline, features: pd.DataFrame, use_log_trick: bool = True
+    model: Pipeline, features: pd.DataFrame
 ) -> np.ndarray:
     predicts = model.predict(features)
-    if use_log_trick:
-        predicts = np.exp(predicts)
     return predicts
 
 
 def evaluate_model(
-    predicts: np.ndarray, target: pd.Series, use_log_trick: bool = False
+    predicts: np.ndarray, target: pd.Series
 ) -> Dict[str, float]:
-    if use_log_trick:
-        target = np.exp(target)
     return {
         "r2_score": r2_score(target, predicts),
         "rmse": mean_squared_error(target, predicts, squared=False),
         "mae": mean_absolute_error(target, predicts),
+        "accuracy": accuracy_score(target, predicts)
     }
