+++ ml_project/saved_models/metrics.json
@@ -1,0 +1,1 @@
+{"r2_score": 0.06971677559912859, "rmse": 0.47907013758040845, "mae": 0.22950819672131148, "accuracy": 0.7704918032786885}
